refactor. working on server.

+++ parser.js
@@ -3,6 +3,8 @@ const path = require('path');
 const $ = require('async');
 const mysql = require('mysql');
 
+const sendEmail = require('./email');
+
 const headers = [ 'Interval', 'DateTime', 'Serial', 'P_AC', 'E_DAY', 'T_WR', 'U_AC', 'U_AC_1', 'U_AC_2', 'U_AC_3', 'I_AC', 'F_AC', 'U_DC_1', 'I_DC_1', 'U_DC_2', 'I_DC_2', 'U_DC_3', 'I_DC_3', 'S', 'E_WR', 'M_MR', 'I_AC_1', 'I_AC_2', 'I_AC_3', 'P_AC_1', 'P_AC_2', 'P_AC_3', 'F_AC_1', 'F_AC_2', 'F_AC_3', 'R_DC', 'PC', 'PCS', 'PCS_LL', 'COS_PHI', 'COS_PHI_LL', 'S_COS_PHI', 'Current_Day_Energy', 'Current_Day_Offset', 'ccEnergyOfDay_WithoutOffset' ];
 const E_WR = headers.indexOf('E_WR');
 const DateTime = headers.indexOf('DateTime');
@@ -10,22 +12,19 @@ const Serial = headers.indexOf('Serial');
 const Current_Day_Energy = headers.indexOf('Current_Day_Energy');
 
 var registry = {};
-var sunrise, sunset;
+var daylights = {};
 
-loadErrors((err) => {
+loadErrorsFromDB((err) => {
   // console.log(errors);
-});
-
-loadDaylightHours(() => {
 });
 
 module.exports = logParser;
 
 function logParser(logfile, done) {
   const self = this;
-  this.db = getDb();
 
   const _done = (err)=> {
+    console.log(err);
     // It's done parsing the `logfile`.
     // Close the DB handle.
     self.db.destroy();
@@ -33,7 +32,16 @@ function logParser(logfile, done) {
     archiveLog(logfile, ()=> { done(err) });
   }
 
-  fs.readFile(logfile, 'utf8', (err, data) => {
+  fs.readFile(logfile, 'ascii', (err, data) => {
+    if(err) {
+      console.log("Error reading file "+logfile);
+      console.log("Most likely it is a directory");
+      // console.log(err);
+      return done();
+    }
+
+    self.db = getDb();
+
     var isLog = false;
     var isHeader = false;
 
@@ -44,8 +52,8 @@ function logParser(logfile, done) {
       return l && isLog;
     });
 
-    // Read each line concurrently and save to DB
-    $.each(lines, processRecord.bind(this), _done);
+    // Read each line and save to DB
+    $.eachSeries(lines, processRecord.bind(this), _done);
   });
 }
 
@@ -63,33 +71,53 @@ function processRecord(r, cb) {
   // If processed record is still not in registry, add it
   registry[serial] || (registry[serial] = { time, currentEnergy });
 
-  // If we're still in daylight hours check
+  // Parse the record and save to DB, then update Devices table
+  $.applyEachSeries([
+    updateInverterData.bind(this),
+    updateDevices.bind(this),
+    notifyProduction.bind(this)
+  ], values, cb);
+  // setTimeout(cb, 10);
+}
+
+function notifyProduction(values, done) {
+  // If time lies in daylight check
   // whether there was power collected during
-  // the last hour.
-  if(isDaylightNow()) {
+  // the last hour, otherwise send an email
+  // notifying possible malfunction
+  const date = new Date(values[DateTime]);
+  const time = date.getTime();
+  const serial = values[Serial];
+  const currentEnergy = parseInt(values[Current_Day_Energy]);
+
+  checkDaylight(date, time, (err, isDaylight) => {
+    if(err) { return done(err) }
+    if(!isDaylight) { return done() }
+
+    // Only if it's daylight
     const lastTime = registry[serial].time;
     // Check every 1 hour
     if(time - lastTime > 1 * 60 * 60 * 1000) {
       const lastEnergy = registry[serial].currentEnergy;
       // If panel collected no power during this
       // time, notify via email
+      console.log('currentEnergy', currentEnergy);
+      console.log('lastEnergy', lastEnergy);
       if(!(currentEnergy > lastEnergy)) {
         sendErrorEmail('999');
       }
       registry[serial] = { time, currentEnergy };
     }
-  }
 
-  // Parse the record and save to DB, then update Devices table
-  $.applyEachSeries([ parseAndSave.bind(this), updateDevices.bind(this) ], values, cb);
-  // setTimeout(cb, 10);
+    done();
+  });
 }
 
 function updateDevices(values, done) {
   const updates = [
-    `UPDATE devices SET LastSeen = ${values[DateTime]} WHERE Serial = ${values[Serial]};`,
-    `UPDATE devices SET LastProduction = ${values[Current_Day_Energy]} WHERE Serial = ${values[Serial]};`,
-    `UPDATE devices SET CurrentError = ${values[E_WR]} WHERE Serial = ${values[Serial]};`
+    `UPDATE devices SET LastSeen = "${values[DateTime]}" WHERE Serial = "${values[Serial]}";`,
+    `UPDATE devices SET LastProduction = ${values[Current_Day_Energy]} WHERE Serial = "${values[Serial]}";`,
+    `UPDATE devices SET CurrentError = ${values[E_WR]} WHERE Serial = "${values[Serial]}";`
   ];
 
   var db = this.db;
@@ -101,14 +129,14 @@ function updateDevices(values, done) {
   }, done);
 }
 
-function parseAndSave(values, cb) {
-
+function updateInverterData(values, cb) {
+  var data = values.slice();
   // Second column is a date time string so we need to quote it
-  values.splice(1, 1, `"${values[DateTime]}"`);
+  data.splice(1, 1, `"${values[DateTime]}"`);
   // Third column `Serial` is a string code so we need to quote it
-  values.splice(2, 1, `"${values[Serial]}"`);
+  data.splice(2, 1, `"${values[Serial]}"`);
   const insert = `INSERT INTO inverterdata (\`${headers.join('\`,\`')}\`, ID)
-    VALUES(${values.join(',')}, NULL);`;
+    VALUES(${data.join(',')}, NULL);`;
 
   this.db.query(insert, (err) => {
     if(err) { console.log(err); return cb(); }
@@ -119,7 +147,7 @@ function parseAndSave(values, cb) {
 // Manage error in solar panel
 var errors = {};
 
-function loadErrors(cb) {
+function loadErrorsFromDB(cb) {
   const db = getDb();
 
   db.query('SELECT ID, Description FROM errorcodes', (err, results) => {
@@ -134,37 +162,63 @@ function loadErrors(cb) {
   });
 }
 
-function loadDaylightHours(cb) {
+function checkDaylight(date, time, cb) {
+  var day = toDayFormat(date);
+
+  // Check if there's daylight already
+  // for the given day in format "DD/MM/YYYY" ...
+  if(daylights[day]) {
+    console.log('sunset', daylights[day].sunset);
+    console.log('time', time);
+    const isDaylight = daylights[day].sunset > time;
+
+    // To keep things async we invoke callback
+    // in the next iteration of the Event Loop
+    return process.nextTick(() => {
+      console.log("isDaylight", isDaylight);
+      cb(null, isDaylight);
+    });
+  }
+
+  // ... otherwise grab from DB
   const db = getDb();
 
-  var today = new Date("2017-02-28");
-  today = today.toISOString().split('T')[0].split('-').reverse().join('/')
-
   const query = `SELECT Dawn, Sunrise, Sunset FROM daylighthours
-    WHERE \`Date\` = '${today}'`;
+    WHERE \`Date\` = '${day}'`;
 
   db.query(query, (err, results) => {
-    var r = results.pop();
-    sunrise = r.Sunrise;
-    sunset = r.Sunset;
+    if(err) {
+      console.log('Error getting daylight hours');
+      console.log(err);
+      return cb(err);
+    }
+
+    var sunrise = '8:00';
+    var sunset = '17:00';
+
+    if(results && results.length) {
+      var r = results.pop();
+      sunrise = r.Sunrise;
+      sunset = r.Sunset;
+    }
 
     const sunriseHours = parseInt(sunrise.split(':').shift());
     const sunriseMinutes = parseInt(sunrise.split(':').pop());
-    var sunriseDate = new Date();
+    var sunriseDate = new Date(time);
     sunriseDate.setHours(sunriseHours);
     sunriseDate.setMinutes(sunriseMinutes);
 
     const sunsetHours = parseInt(sunset.split(':').shift());
     const sunsetMinutes = parseInt(sunset.split(':').pop());
-    var sunsetDate = new Date();
+    var sunsetDate = new Date(time);
     sunsetDate.setHours(sunsetHours);
     sunsetDate.setMinutes(sunsetMinutes);
 
-    sunrise = sunriseDate.getTime();
-    sunset = sunsetDate.getTime();
+    daylights[day] = { sunrise: sunriseDate.getTime(), sunset: sunsetDate.getTime() };
 
     db.destroy();
-    cb();
+
+    checkDaylight(date, time, cb)
   });
 }
 
@@ -181,34 +235,6 @@ function archiveLog(logfile, cb) {
   });
 }
 
-// Email facilities
-const nodemailer = require('nodemailer');
-const transporter = nodemailer.createTransport({
-  sendmail: true,
-  newline: 'unix',
-  path: '/usr/sbin/sendmail'
-});
-
-const errorEmailFrom = 'sender@address';
-const errorEmailTo = 'receiver@address';
-const errorEmailSubject = 'An error occurred in panel operation #PANEL_ID';
-
-function sendErrorEmail(errorCode) {
-  const code = parseInt(errorCode);
-  code && console.log('sendErrorEmail', errors[errorCode]);
-  transporter.sendMail({
-     from: errorEmailFrom,
-     to: errorEmailFrom,
-     subject: errorEmailSubject,
-     html: `<b>${errorEmailSubject}</b>`,
-     text: errorEmailSubject
-  }, (err, info) => {
-    if(err) { console.log(`Error sending email: ${err}`); }
-    console.log(info.envelope);
-    console.log(info.messageId);
-  });
-}
-
 function getDb() {
   return mysql.createConnection({
     host     : 'localhost',
@@ -218,6 +244,26 @@ function getDb() {
   });
 }
 
-function isDaylightNow() {
-  return Date.now() < sunset;
+function toDayFormat(date) {
+  return date.toISOString().split('T')[0].split('-').reverse().join('/');
+}
+
+function sendErrorEmail(errorCode) {
+  const errorEmailFrom = 'sender@address';
+  const errorEmailTo = 'receiver@address';
+  const errorEmailSubject = 'An error occurred in panel operation #PANEL_ID';
+
+  const code = parseInt(errorCode);
+  if(code) {
+    var err = errors[errorCode];
+    sendEmail({
+       from: errorEmailFrom,
+       to: errorEmailFrom,
+       subject: errorEmailSubject,
+       html: `<b>${err}</b>`,
+       text: err
+    });
+
+    console.log('sendErrorEmail', err);
+  }
 }
